// src/lib/api/generated/schemas/createInitialQuizVersionDtoDifficulty.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Quiz API
 * REST API for the quiz application
 * OpenAPI spec version: 1.0
 */

/**
 * Quiz difficulty level
 */
export type CreateInitialQuizVersionDtoDifficulty = typeof CreateInitialQuizVersionDtoDifficulty[keyof typeof CreateInitialQuizVersionDtoDifficulty];


 
export const CreateInitialQuizVersionDtoDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;